app/requirements.txt
fastapi==0.95.1
python-multipart==0.0.6
GitPython==3.1.31
uvicorn[standard]==0.22.0
